# backend/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OpenISMS API",
    "description": "REST API für das Information Security Management System (OpenISMS).\n\n## Authentifizierung\nAlle geschützten Endpunkte erfordern einen **Bearer JWT-Token** im `Authorization`-Header:\n```\nAuthorization: Bearer <token>\n```\nDen Token erhalten Sie über `POST /api/auth/login`.\n\n## Rollen\n- `admin` – Vollzugriff\n- `assessor` – Risikobewerter / Auditor\n- `it-staff` – IT-Mitarbeiter (Security-Daten)\n- `dpo` – Datenschutzbeauftragter\n- `owner` – Asset-Verantwortlicher\n- `management` – Nur-Lesen (Management-Ebene)\n- `employee` – Schulungs-Tab, eigene Trainingsübersicht\n- `viewer` – Nur Lesezugriff\n\n**Custom Roles**: Eigene Rollennamen mit Basisrolle über `/api/admin/custom-roles` verwalten.",
    "version": "2.2.0",
    "contact": { "name": "ISMS Support" }
  },
  "servers": [
    { "url": "/api", "description": "Aktuelle Instanz" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": { "error": { "type": "string" } }
      },
...
      "Asset": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "type": { "type": "string", "enum": ["hardware","software","information","process","service","personal","application","data","ai_application","ai_agent","other"] },
          "description": { "type": "string" },
          "classification": { "type": "string", "enum": ["public","internal","confidential","secret"] },
          "hosting_type": { "type": "string", "enum": ["on-premise","cloud_public","cloud_private","hybrid"] },
          "lifecycle_status": { "type": "string", "enum": ["evaluation","production","maintenance","archived"] },
          "status": { "type": "string", "enum": ["active","inactive","decommissioned"] },
          "nis2_relevant": { "type": "boolean" },
          "owner_id": { "type": "integer" },
          "assessor_id": { "type": "integer" },
          "vendor_id": { "type": "integer", "nullable": true },
          "patch_status": { "type": "string", "enum": ["up-to-date","pending","critical"] },
          "rto_hours": { "type": "integer", "nullable": true },
          "rpo_hours": { "type": "integer", "nullable": true },
          "cve_critical": { "type": "integer" },
          "cve_high": { "type": "integer" },
          "cve_medium": { "type": "integer" },
          "cve_low": { "type": "integer" },
          "frameworks": { "type": "array", "items": { "type": "string" } },
          "parent_id": { "type": "integer", "nullable": true },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/assets": {
      "get": {
        "tags": ["Assets"],
        "summary": "Alle Assets abrufen",
        "parameters": [
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "type", "in": "query", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "hosting_type", "in": "query", "schema": { "type": "string" } },
          { "name": "classification", "in": "query", "schema": { "type": "string" } },
          { "name": "lifecycle_status", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Asset-Liste", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Asset" } } } } } }
      },
      "post": {
        "tags": ["Assets"],
        "summary": "Neues Asset anlegen",
        "description": "Erfordert Rolle: admin, assessor, it-staff, dpo",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Asset" } } } },
        "responses": {
          "201": { "description": "Asset erstellt", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Asset" } } } },
          "400": { "description": "Validierungsfehler", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/assets/{id}": {
      "get": {
        "tags": ["Assets"],
        "summary": "Asset-Details abrufen",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Asset mit verknüpften Daten" }, "404": { "description": "Nicht gefunden" } }
      },
      "put": {
        "tags": ["Assets"],
        "summary": "Asset aktualisieren",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Asset" } } } },
        "responses": { "200": { "description": "Asset aktualisiert" } }
      },
      "delete": {
        "tags": ["Assets"],
        "summary": "Asset löschen",
        "description": "Erfordert Rolle: admin",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Asset gelöscht" } }
      }
    },
...
    "/import/assets": {
      "post": {
        "tags": ["Import"],
        "summary": "Assets per CSV oder Excel importieren",
        "description": "Erfordert Rolle: admin, assessor. Unterstützt CSV und .xlsx.",
        "requestBody": { "required": true, "content": { "multipart/form-data": { "schema": { "type": "object", "properties": { "file": { "type": "string", "format": "binary" } } } } } },
        "responses": { "200": { "description": "Import-Ergebnis mit Erfolg/Fehler-Zählung" } }
      }
    },
...
  },
...
}